% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{summarize_missing}
\alias{summarize_missing}
\title{Summarize Missing Values in a Data Frame}
\usage{
summarize_missing(.data)
}
\arguments{
\item{.data}{A data frame.}
}
\value{
A data frame summarizing the missing values in each column.
The data frame has three columns: \code{column} (the name of the column),
\code{n_missing} (the number of NAs), and \code{pct_missing} (the percentage of NAs).
}
\description{
This function scans each column of a data frame and returns a summary
data frame with the count and percentage of missing (\code{NA}) values for each one.
}
\examples{
# To use the example, you must first load the data from the package
# utils::data("messy_data", package = "deskcleanerV2")

summarize_missing(messy_data)
}
